Tuning Checkpoint
Checkpoint with different policies and tuning mechanisms. The next
checkpoint will have only one mechanism (semaphore).

+++ src/main/java/jvstm/tuning/ThreadStatistics.java
@@ -1,86 +1,108 @@
 package jvstm.tuning;
 
-public class ThreadStatistics {
+public class ThreadStatistics
+{
 
-	private volatile int transactionCount;
-	private volatile int commitCount;
-	private volatile int abortCount;
+	private volatile long transactionCount;
+	private volatile long commitCount;
+	private volatile long abortCount;
 	private volatile long threadId;
-	private volatile int nestingLevel;
-	
-	public ThreadStatistics(long threadId, int nestingLevel) {
+	private volatile long nestingLevel;
+
+	public ThreadStatistics(long threadId, int nestingLevel)
+	{
 		this.threadId = threadId;
 		this.nestingLevel = nestingLevel;
 	}
-	
-	public void reset() {
+
+	public void reset()
+	{
 		transactionCount = 0;
 		commitCount = 0;
 		abortCount = 0;
 		nestingLevel = 0;
 	}
-	
-	public void reset(long threadId) {
+
+	public void reset(long threadId)
+	{
 		this.threadId = threadId;
 		reset();
 	}
-	
-	
-	//to do: use lock?
-	//does not cause memory consistency errors but is not thread-safe
-	public void addTo(ThreadStatistics stat) {
+
+	// to do: use lock?
+	// does not cause memory consistency errors but is not thread-safe
+	public void addTo(ThreadStatistics stat)
+	{
 		stat.transactionCount += this.transactionCount;
 		stat.commitCount += this.commitCount;
 		stat.abortCount += this.abortCount;
 	}
-	
-	public long getThreadId() {
+
+	public long getThreadId()
+	{
 		return threadId;
 	}
-	
-	public int getNestingLevel() {
+
+	public long getNestingLevel()
+	{
 		return nestingLevel;
 	}
 
-	protected void setNestingLevel(int nestingLevel) {
+	protected void setNestingLevel(int nestingLevel)
+	{
 		this.nestingLevel = nestingLevel;
 	}
 
-	public int getTransactionCount() {
+	public long getTransactionCount()
+	{
 		return transactionCount;
 	}
-	public void setTransactionCount(int transactionCount) {
+
+	public void setTransactionCount(int transactionCount)
+	{
 		this.transactionCount = transactionCount;
 	}
-	public void incTransactionCount() {
+
+	public void incTransactionCount()
+	{
 		transactionCount++;
 	}
-	
-	public int getCommitCount() {
+
+	public long getCommitCount()
+	{
 		return commitCount;
 	}
-	public void setCommitCount(int commitCount) {
+
+	public void setCommitCount(int commitCount)
+	{
 		this.commitCount = commitCount;
 	}
-	public void incCommitCount() {
+
+	public void incCommitCount()
+	{
 		commitCount++;
 	}
-	
-	public int getAbortCount() {
+
+	public long getAbortCount()
+	{
 		return abortCount;
 	}
-	public void setAbortCount(int abortCount) {
+
+	public void setAbortCount(int abortCount)
+	{
 		this.abortCount = abortCount;
 	}
-	public void incAbortCount() {
+
+	public void incAbortCount()
+	{
 		abortCount++;
 	}
 
 	@Override
-	public String toString() {
+	public String toString()
+	{
 		return "ThreadStatistics [transactionCount=" + transactionCount + ", commitCount=" + commitCount
 				+ ", abortCount=" + abortCount + ", threadId=" + threadId + ", nestingLevel=" + nestingLevel + "]";
 	}
-	
-	
+
 }
